add macOS flag

+++ _xtoa.c
@@ -7,7 +7,7 @@
 #include "_xtoa.h"
 
 /* Converts 32-bit unsigned integer to a buffer. Base is 16 by default. */
-void _utoa(u32 n, char* buf, int next, u8 base) {
+void _utoa(u32 n, char* buf, i8 next, u8 base) {
     u32  r, f;
     bool foundNonzero = false;
